neo-core2/src/crypto/keys/private_key.rs
extern crate secp256k1;
extern crate sha2;
extern crate hex;
extern crate rand;
extern crate num_bigint;
extern crate num_traits;
extern crate openssl;

use secp256k1::{Secp256k1, SecretKey, PublicKey, Message, Signature};
use sha2::{Sha256, Digest};
use hex::decode;
use rand::rngs::OsRng;
use num_bigint::BigInt;
use num_traits::Num;
use openssl::bn::BigNum;
use openssl::ec::{EcKey, EcGroup};
use openssl::nid::Nid;
use std::fmt;

pub struct PrivateKey {
    key: SecretKey,
}

impl PrivateKey {
    pub fn new() -> Result<PrivateKey, secp256k1::Error> {
        let secp = Secp256k1::new();
        let (secret_key, _) = secp.generate_keypair(&mut OsRng);
        Ok(PrivateKey { key: secret_key })
    }

    pub fn new_secp256k1() -> Result<PrivateKey, secp256k1::Error> {
        let secp = Secp256k1::new();
        let (secret_key, _) = secp.generate_keypair(&mut OsRng);
        Ok(PrivateKey { key: secret_key })
    }

    pub fn from_hex(hex_str: &str) -> Result<PrivateKey, hex::FromHexError> {
        let bytes = decode(hex_str)?;
        PrivateKey::from_bytes(&bytes)
    }

    pub fn from_bytes(bytes: &[u8]) -> Result<PrivateKey, secp256k1::Error> {
        let secret_key = SecretKey::from_slice(bytes)?;
        Ok(PrivateKey { key: secret_key })
    }

    pub fn from_asn1(asn1: &[u8]) -> Result<PrivateKey, openssl::error::ErrorStack> {
        let ec_key = EcKey::private_key_from_der(asn1)?;
        let group = EcGroup::from_curve_name(Nid::SECP256K1)?;
        let private_key = ec_key.private_key().to_vec();
        let secret_key = SecretKey::from_slice(&private_key)?;
        Ok(PrivateKey { key: secret_key })
    }

    pub fn public_key(&self) -> PublicKey {
        let secp = Secp256k1::new();
        PublicKey::from_secret_key(&secp, &self.key)
    }

    pub fn from_wif(wif: &str) -> Result<PrivateKey, secp256k1::Error> {
        // Implement WIF decoding here
        unimplemented!()
    }

    pub fn wif(&self) -> String {
        // Implement WIF encoding here
        unimplemented!()
    }

    pub fn destroy(&mut self) {
        // Clear the private key from memory
        self.key = SecretKey::from_slice(&[0u8; 32]).unwrap();
    }

    pub fn address(&self) -> String {
        // Implement address derivation here
        unimplemented!()
    }

    pub fn get_script_hash(&self) -> [u8; 20] {
        // Implement script hash derivation here
        unimplemented!()
    }

    pub fn sign(&self, data: &[u8]) -> Result<Signature, secp256k1::Error> {
        let secp = Secp256k1::new();
        let message = Message::from_slice(&Sha256::digest(data))?;
        secp.sign(&message, &self.key)
    }

    pub fn sign_hash(&self, hash: &[u8]) -> Result<Signature, secp256k1::Error> {
        let secp = Secp256k1::new();
        let message = Message::from_slice(hash)?;
        secp.sign(&message, &self.key)
    }

    pub fn sign_hashable(&self, net: u32, hh: &[u8]) -> Result<Signature, secp256k1::Error> {
        let secp = Secp256k1::new();
        let mut hasher = Sha256::new();
        hasher.update(&net.to_be_bytes());
        hasher.update(hh);
        let hash = hasher.finalize();
        let message = Message::from_slice(&hash)?;
        secp.sign(&message, &self.key)
    }

    pub fn bytes(&self) -> Vec<u8> {
        self.key[..].to_vec()
    }
}

impl fmt::Display for PrivateKey {
    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
        write!(f, "{}", hex::encode(self.bytes()))
    }
}

fn clear(bytes: &mut [u8]) {
    for byte in bytes.iter_mut() {
        *byte = 0;
    }
}
